+++ .github/workflows/test.yml
@@ -22,4 +22,4 @@ jobs:
           go-version: ${{ matrix.go-version }}
       - uses: actions/checkout@v3
       - name: Run the tests
-        run: go test -tags=whitebox,blackbox ./...
+        run: go test ./...
