+++ blackbox_test.go
@@ -1,5 +1,3 @@
-//go:build blackbox
-
 package evs_test
 
 import (
@@ -10,7 +8,7 @@ import (
 	"github.com/thenorthnate/evs"
 )
 
-func TestFrom(t *testing.T) {
+func TestBlackBoxFrom(t *testing.T) {
 	err := errors.New("Hello, world")
 	newErr := evs.From(err).Err()
 	if !strings.Contains(newErr.Error(), "*evs.Error[github.com/thenorthnate/evs.Std]: Hello, world") {
